// src/modules/book/book.enum.ts
export enum Visibility {
    PUBLISH = 'publish',
    DELETE = 'delete',
}

export enum ContentType {
    PRESENTATION = 'presentation',
    TEXT = 'text',
    VIDEO = 'video',
    AUDIO = 'audio',
}

export enum PaymentMethod {
    Credit = 'credit',
    Point = 'point',
}

export enum PointEnabled {
    ACTIVE = 'active',
    INACTIVE = 'inactive',
}